requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.1
flatbuffers==25.9.23
fonttools==4.61.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jax==0.6.2
jaxlib==0.6.2
jiter==0.12.0
kiwisolver==1.4.9
matplotlib==3.10.7
mediapipe==0.10.21
ml_dtypes==0.5.4
numpy==1.26.4
openai==2.9.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==12.0.0
protobuf==4.25.8
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
